create DataTable component and delete VendorsTable
